refactor(bench-python): split the launcher's two longest key checks into named refusals
`bench/harbor_adapter/stella_harbor/secure_launcher.py` is the file the #4392
audit named as the worst complexity case outside `docs`, and the cause is
functions long enough that nobody holds one in their head. The launcher
decides whether a benchmark trial is claim-eligible; a function nobody can
read is a function nobody can audit.

Measured first, so the split targets the real offenders rather than the ones
that look long. The two worst dedicated-key checks were 182 and 132 lines.
They are now thirteen functions, none over 80, in two new modules:

- `value_shapes.py` — the value-shape refusals every validator here shares:
  `aware_timestamp`, `finite_nonnegative_number`, `require_exact_object`,
  `utc_text`, `parse_github_timestamp`. Pure, clock-free, network-free, so a
  caller can test one refusal without building a launch.
- `provider_key.py` — the dedicated OpenRouter key: `prior_stage_outcome`
  reads the public ledger and refuses unless every paid stage bound this one
  key and reconciled its spend; `validate_live_provider_key` reads its live
  state and refuses unless it still matches the intent and can cover the job.
  Each refusal is now a named predicate.

The layering is `value_shapes` (leaf) <- `provider_key` <- `secure_launcher`,
which is what avoids the import cycle a partial extraction would have had.
`secure_launcher` imports the moved names back under the private names it
already used, so every call site and every `monkeypatch.setattr` target is
unchanged.

Every refusal message is byte-identical, and no test was edited.
`_FIXED_ADAPTER_SOURCE_PATHS` gains both modules, which is that enumeration's
own documented rule — an adapter source not listed there would be executed
without being byte-compared to its published source.

`secure_launcher.py` falls 4377 -> 4010 lines.
`scripts/file-size-baseline.txt` is deliberately NOT retightened: AGENTS.md
makes a retighten its own PR, and #4894 needs the headroom that row already
has. `PRIOR_STAGE_OUTCOME_FIELDS` and `PROVIDER_KEY_LIVE_SNAPSHOT_FIELDS`,
#4894's subject, stay in `secure_launcher.py` where its guard looks for them;
neither moved function referenced either, which is #4894's finding.

Refs #4759
Refs #4392
